requirements.txt
PySide6
llama-cpp-python
requests
Pillow
